Changes from Stream (twitch.tv/secretspast)
-x64 MatchServer/Gunz exe now builds.
-Two configurations exist now; Debug/Release. Cleaned up unneeded project configurations.
-Cleaned up a shit ton of warnings, mostly in gamemodes and zitem parsing code.
-New Gitignore file to exclude built libraries.
-VS2019 support w/ latest Windows SDK.

+++ Stable/CSCommon/Include/AntiSql.h
@@ -1,8 +1,8 @@
 inline bool SqlCheckString(char *str)
 {
 	string szString = str;
-	int len = szString.length();
-	for(int i = 0 ; i < len ; i++)
+	size_t len = szString.length();
+	for(size_t i = 0 ; i < len ; i++)
 	{
 		if(szString[i] == '\'')
 		{
@@ -28,8 +28,8 @@ inline void CleanTSQLString(string& szString)
 inline bool StringCheckFormat(char *str)
 {
 	string szString = str;
-	int len = szString.length();
-	for(int i = 0 ; i < len ; i++)
+	size_t len = szString.length();
+	for(size_t i = 0 ; i < len ; i++)
 	{
 		if(szString[i] == '%')
 		{
@@ -42,8 +42,8 @@ inline bool StringCheckFormat(char *str)
 inline bool StringCheckName(char *str)
 {
 	string szString = str;
-	int len = szString.length();
-	for(int i = 0 ; i < len ; i++)
+	size_t len = szString.length();
+	for(size_t i = 0 ; i < len ; i++)
 	{
 		if(szString[i] == '^' || szString[i] == '\t' || szString[i] == '\n')
 		{
